+++ src/F1.Infrastructure/Data/Entities/MigrationImportRaceDiffEntity.cs
@@ -7,8 +7,8 @@ public sealed class MigrationImportRaceDiffEntity
     public string RaceCode { get; set; } = string.Empty;
     public string Subject { get; set; } = string.Empty;
     public int ImportedPoints { get; set; }
-    public int CalculatedPoints { get; set; }
-    public int DeltaPoints { get; set; }
+    public decimal CalculatedPoints { get; set; }
+    public decimal DeltaPoints { get; set; }
     public string ReasonCode { get; set; } = string.Empty;
     public bool IsExpectedVariance { get; set; }
     public string? ExpectedVarianceReasonCode { get; set; }
